scheduler/requirements.txt
PySide6
psycopg2-binary
deap
ortools